+++ config.example.js
@@ -56,7 +56,7 @@ const exampleConfig = {
     emailUsername: false,
     // Also return the username and UUID when creating a session
     sendNameAndUUID: false,
-    // If a number is set here, the token for password reset will be shortened to that length (e.g. 8)
+    // If a number is set here, the token for password reset will be shortened to that length (at least 12)
     tokenLengthOnReset: false,
     // Custom names for the username and password fields in your sign-in form
     usernameField: 'user',